Module private_currency::api

HTTP API for the service.

Structs

CheckedWalletProof
Information about wallet state useful for a client, obtained after checking a WalletProof.

TrustAnchor
Trust anchor for block verification.

WalletProof
Cryptographically authenticated proof of the state for a single wallet.

WalletQuery
Query for the wallet endpoint.

Enums

Api
HTTP API for the private cryptocurrency service.

BlockVerifyError
Error occuring during block header verification.

FullEvent
Event changing balance of a wallet.

ProofDescription
Description of a part of a WalletProof.

VerifyError
Error during WalletProof verification.