Struct private_currency::crypto::Commitment

pub struct Commitment { /* fields omitted */ }

Pedersen commitment to an integer value.

Theory

A Pedersen commitment has the form (in the additive notation)

Comm(x; r) = xG + rH,

where

G and H are two generators in a prime-order group Q, with unknown discrete logarithm relationships among them (i.e., nobody knows k such as G = kH)

x is the committed value (it’s a residue class modulo the group order |Q|, but we may essentially treat it as an integer)

r is the blinding factor (also a residue class modulo the group order |Q|). Usually, it’s chosen randomly from a cryptographically secure RNG.

Q, G and H are public parameters of the scheme shared among all commitments, while x and r are private. (x, r) is called an opening to the commitment; knowing the opening, it’s easy to check if it corresponds to the given commitment.

Under common security assumptions, a Pedersen commitment is

perfectly hiding (a party not knowing x and r cannot find them out from Comm(..)), and

computationally binding (a party cannot produce (x', r') != (x, r) such that they will open to the same commitment).

Commitment arithmetic

It is possible to add and subtract Pedersen commitments; the result is a commitment to the sum / difference of corresponding values. This fact is what allows using commitments in private currency.

Implementation details

We use a Ristretto group built on top of Curve25519 as Q. Generators G and H are constructed according to the default scheme in the bulletproofs implementation.

Examples

let (mut commitment, mut opening) = Commitment::new(42);
assert_eq!(opening.value, 42);
assert_eq!(commitment, Commitment::from_opening(&opening));

let (other_commitment, other_opening) = Commitment::new(23);
commitment -= other_commitment;
opening -= other_opening;
assert_eq!(opening.value, 19);
assert_eq!(commitment, Commitment::from_opening(&opening));

Methods

impl Commitment

pub fn new(value: u64) -> (Self, Opening)

Creates a commitment with a randomly chosen blinding.

Return value

Returns the created commitment and the corresponding opening for it.

pub fn from_opening(opening: &Opening) -> Self

Creates a commitment from the given opening.

pub fn with_no_blinding(value: u64) -> Self

Creates a commitment with no blinding factor.

Warning. The commitments created in this way are not hiding. Use them only if you know what you’re doing.

pub fn from_slice(slice: &[u8]) -> Option<Self>

Attempts to deserialize a commitment from byte slice.

pub fn to_bytes(&self) -> Vec<u8>

Serializes this commitment to bytes.

Implementation details

The commitment is serialized as a single compressed Ristretto point (i.e., 32 bytes).

pub fn verify(&self, opening: &Opening) -> bool

Verifies if this commitment corresponds to the provided opening.