Struct private_currency::storage::Wallet

pub struct Wallet { /* fields omitted */ }

Wallet summary.

Methods

impl Wallet

pub fn new(
public_key: &PublicKey,
balance: Commitment,
history_len: u64,
last_send_index: u64,
history_hash: &Hash,
unaccepted_transfers_hash: &Hash
) -> Wallet

Creates a new instance with given parameters.

pub fn public_key(&self) -> &PublicKey

Ed25519 public key associated with the wallet. Transactions originating from the wallet need to be digitally signed with the paired secret key.

pub fn balance(&self) -> Commitment

Commitment to the current wallet balance.

pub fn history_len(&self) -> u64

Number of entries in the wallet history.

pub fn last_send_index(&self) -> u64

Index of the last outgoing transfer in the wallet history.

pub fn history_hash(&self) -> &Hash

Merkle root of the wallet history list.

pub fn unaccepted_transfers_hash(&self) -> &Hash

Merkle root of the unaccepted incoming transfers.

impl Wallet

pub fn info(&self) -> WalletInfo

Retrieves the wallet summary.

pub fn encryption_key(&self) -> PublicKey

Computes the encryption key associated with the wallet.