Struct private_currency::transactions::CreateWallet

pub struct CreateWallet { /* fields omitted */ }

Transaction for creating a new wallet.

Notes

This transaction specifies only the Ed25519 verification key used to check digital signatures of transactions authored by the wallet owner. The public encryption key of the wallet owner is deterministically derived from the verification key.

Methods

impl CreateWallet

pub fn new(key: &PublicKey, secret_key: &SecretKey) -> CreateWallet

Creates message and signs it.

pub fn new_with_signature(
key: &PublicKey,
signature: &Signature
) -> CreateWallet

Creates message and appends existing signature.

pub fn to_hex(&self) -> String

Returns the hex representation of the binary data. Lower case letters are used (e.g. f9b4ca).

pub fn key(&self) -> &PublicKey

Ed25519 key for the wallet.